Eliminate ET_Generic, and ext.stub
