+++ config/splat.us.stno3.yaml
@@ -67,7 +67,8 @@ segments:
       - [0x3AF0, data]
       - [0x3B30, data]
       - [0x3B50, .data, e_bone_scimitar]
-      - [0x3C44, data]
+      - [0x3C44, .data, e_bat]
+      - [0x3C7C, data]
       - [0x3CC4, .data, rooms]
       - [0x3DC8, .data, e_layout] # layout entries data
       - [0x4CE0, .data, cutscene_data]
@@ -104,7 +105,7 @@ segments:
       - [0x37738, .rodata, e_merman2]
       - [0x3775C, .rodata, 53F38]
       - [0x37784, .rodata, e_bone_scimitar]
-      - [0x377A4, .rodata, 56264] # EntityBat
+      - [0x377A4, .rodata, e_bat] # EntityBat
       - [0x377B8, .rodata, 564B0] # EntityZombie
       - [0x377CC, .rodata, prim_helpers]
       - [0x377D4, c]
@@ -129,7 +130,7 @@ segments:
       - [0x516B4, c, e_merman2]
       - [0x53F38, c]
       - [0x559d0, c, e_bone_scimitar]
-      - [0x56264, c]
+      - [0x56264, c, e_bat]
       - [0x564B0, c]
       - [0x56880, c, prim_helpers]
       - [0x570B0, .bss, create_entity]
