+++ config/symbols.stmad.txt
@@ -5,6 +5,7 @@ MAD_pStObjLayoutVertical = 0x801803C8;
 MAD_EntityUpdates = 0x8018049C;
 g_EInitBreakable = 0x801804F0;
 g_EInitUnkId13 = 0x8018052C;
+g_EInitUnkId12 = 0x80180538;
 g_EInitCommon = 0x80180544;
 g_EInitDamageNum = 0x80180550;
 g_eBreakableAnimations = 0x80180644;
@@ -36,6 +37,7 @@ g_pStrObjLayout18 = 0x80181AF8;
 g_pStrObjLayout19 = 0x80181BA2;
 g_pStrObjLayout20 = 0x80181C4C;
 g_pStrObjLayout21 = 0x80181CF6;
+EntityUnkId12 = 0x8018E674;
 EntityBreakable = 0x8018E830;
 EntityDamageDisplay = 0x8018FEA0;
 EntityRedDoor = 0x80191120;
