+++ config/symbols.us.stdre.txt
@@ -5,6 +5,7 @@ g_EInitObtainable = 0x80180464;
 g_EInitParticle = 0x80180470;
 g_EInitInteractable = 0x80180488;
 g_EInitUnkId13 = 0x80180494;
+g_EInitUnkId12 = 0x801804A0;
 g_EInitCommon = 0x801804AC;
 g_EInitDamageNum = 0x801804B8;
 DRE_RedDoorTiles = 0x80180550;
@@ -19,6 +20,7 @@ g_ESoulStealOrbAngles = 0x8018138C;
 g_ESoulStealOrbSprt = 0x8018139C;
 g_ESoulStealOrbAnim = 0x801813FC;
 EntityUnkId11 = 0x80191A64;
+EntityUnkId12 = 0x80191B44;
 EntityBreakable = 0x80191D00;
 EntityBackgroundClouds = 0x80191E34;
 EntitySuccubus = 0x80192104;
