+++ config/symbols.us.stno3.txt
@@ -13,6 +13,7 @@ g_EInitWaterObject = 0x80180B48;
 g_EInitWaterSplash = 0x80180B54;
 g_EInitBoneScimitar = 0x80180B78;
 g_EInitScimitarParts = 0x80180B84;
+g_EInitBat = 0x80180B90;
 NO3_RedDoorTiles = 0x80180DA0;
 g_eBreakableAnimations = 0x80180E50;
 g_eBreakableHitboxes = 0x80180E70;
@@ -101,7 +102,7 @@ NO3_EntityCutscene = 0x801BEEB0;
 EntityRoomTransition2 = 0x801BFF00;
 EntityDeathStolenItem = 0x801C033C;
 EntityDeath = 0x801C083C;
-EntityUnkId5B = 0x801C11A4;
+EntityDeathScythe = 0x801C11A4;
 EntityDeathScytheShadow = 0x801C12E8;
 CreateExplosionPuff = 0x801C13F8;
 EntityWargExplosionPuffOpaque = 0x801C14B8;
