+++ src/st/dre/e_breakable.c
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "dre.h"
+#include "sfx.h"
+
+extern u16 g_EInitBreakable[];
+extern u8* g_eBreakableAnimations[];
+extern u8 g_eBreakableHitboxes[];
+extern u8 g_eBreakableExplosionTypes[];
+extern u16 g_eBreakableanimSets[];
+extern u8 g_eBreakableDrawModes[];
+
+#include "../e_breakable.h"
