+++ src/st/entity_unk_id12.h
@@ -18,20 +18,20 @@ static u16 g_EntityUnkId12TilemapProps[] = {
     0x0000, 0x0000, 0x0500, 0x0400, 0x0000, 0x0000, 0x0500, 0x0400, 0x0000,
     0x01FC, 0x0500, 0x02FC, 0x00F0, 0x01FC, 0x0310, 0x02FC, 0x00F0, 0x01FC,
     0x0310, 0x02FC,
-#if defined(STAGE_IS_NO3)
+#if defined(STAGE_IS_NO3) || defined(STAGE_IS_DRE)
     0x0104, 0x0204, 0x0000, 0x0000, 0x0004, 0x0004, 0x0000, 0x0000
 #endif
 };
 
 extern u16 g_EInitUnkId12[];
-void EntityUnkId12(Entity* entity) {
+void EntityUnkId12(Entity* self) {
     u16 var_s0;
     u16 params;
     u16* tilemapProps;
 
-    entity->unk6D[0] = 0;
-    params = entity->params;
-    if (entity->step) {
+    self->unk6D[0] = 0;
+    params = self->params;
+    if (self->step) {
         switch (params) {
         case 4:
         case 5:
@@ -43,15 +43,15 @@ void EntityUnkId12(Entity* entity) {
             if (g_pads->pressed & PAD_TRIANGLE) {
                 g_Tilemap.x = 0;
                 g_Tilemap.width = 1280;
-                entity->step++;
+                self->step++;
                 return;
             }
             break;
         }
 
-        if (entity->hitParams) {
+        if (self->hitParams) {
             var_s0 = GetSideToPlayer();
-            if (entity->ext.generic.unk7C.u) {
+            if (self->ext.ent12.unk7C) {
                 var_s0 &= 2;
                 var_s0 *= 2;
             } else {
@@ -67,13 +67,13 @@ void EntityUnkId12(Entity* entity) {
         }
     } else {
         InitializeEntity(g_EInitUnkId12);
-        var_s0 = entity->ext.generic.unk7C.u = g_EntityUnkId12Data[params];
+        var_s0 = self->ext.ent12.unk7C = g_EntityUnkId12Data[params];
         if (var_s0) {
-            entity->hitboxWidth = g_EntityUnkId12Hitbox[params];
-            entity->hitboxHeight = 16;
+            self->hitboxWidth = g_EntityUnkId12Hitbox[params];
+            self->hitboxHeight = 16;
         } else {
-            entity->hitboxWidth = 16;
-            entity->hitboxHeight = g_EntityUnkId12Hitbox[params];
+            self->hitboxWidth = 16;
+            self->hitboxHeight = g_EntityUnkId12Hitbox[params];
         }
     }
 }
