+++ src/st/mad/D8C8.c
@@ -92,100 +92,3 @@ void func_8018E13C(u16* tilemap) {
 
 void func_8018E1D4(void);
 INCLUDE_ASM("asm/us/st/mad/nonmatchings/D8C8", func_8018E1D4);
-
-void func_8018E5AC(Entity* self) {
-    s32 temp_v0;
-    ObjInit* objInit = &D_8018056C[self->params];
-
-    if (self->step == 0) {
-        InitializeEntity(g_EInitCommon);
-        self->animSet = objInit->animSet;
-        self->zPriority = objInit->zPriority;
-        self->unk5A = objInit->unk5A;
-        self->palette = objInit->palette;
-        self->drawFlags = objInit->drawFlags;
-        self->drawMode = objInit->drawMode;
-        temp_v0 = objInit->unkC;
-        if (temp_v0 != 0) {
-            self->flags = temp_v0;
-        }
-    }
-    AnimateEntity(objInit->animFrames, self);
-}
-
-void func_8018E674(Entity* self) {
-    u16 var_s0;
-    u16 params;
-    u16* tilemapProps;
-
-    self->unk6D[0] = 0;
-    params = self->params;
-    if (self->step) {
-        switch (params) {
-        case 4:
-        case 5:
-            if (g_Tilemap.x) {
-                return;
-            }
-            break;
-        case 6:
-            if (g_pads->pressed & PAD_TRIANGLE) {
-                g_Tilemap.x = 0;
-                g_Tilemap.width = 1280;
-                self->step++;
-                return;
-            }
-            break;
-        }
-        if (self->hitParams) {
-            var_s0 = GetSideToPlayer();
-            if (self->ext.generic.unk7C.u) {
-                var_s0 &= 2;
-                var_s0 *= 2;
-            } else {
-                var_s0 &= 1;
-                var_s0 *= 4;
-            }
-            params = (params << 3) + var_s0;
-            tilemapProps = &D_801805C4[params];
-            g_Tilemap.x = *tilemapProps++;
-            g_Tilemap.y = *tilemapProps++;
-            g_Tilemap.width = *tilemapProps++;
-            g_Tilemap.height = *tilemapProps;
-        }
-    } else {
-        InitializeEntity(D_80180538);
-        var_s0 = self->ext.timer.t = D_801805BC[params];
-        if (var_s0) {
-            self->hitboxWidth = D_801805B4[params];
-            self->hitboxHeight = 16;
-            return;
-        }
-        self->hitboxWidth = 16;
-        self->hitboxHeight = D_801805B4[params];
-    }
-}
-
-void EntityBreakable(Entity* entity) {
-    u16 breakableType = entity->params >> 0xC;
-    if (entity->step) {
-        AnimateEntity(g_eBreakableAnimations[breakableType], entity);
-        if (entity->hitParams) { // If the candle is destroyed
-            Entity* entityDropItem;
-            g_api.PlaySfx(SFX_CANDLE_HIT);
-            entityDropItem = AllocEntity(&g_Entities[224], &g_Entities[256]);
-            if (entityDropItem != NULL) {
-                CreateEntityFromCurrentEntity(E_EXPLOSION, entityDropItem);
-                entityDropItem->params =
-                    g_eBreakableExplosionTypes[breakableType];
-            }
-            ReplaceBreakableWithItemDrop(entity);
-        }
-    } else {
-        InitializeEntity(g_EInitBreakable);
-        entity->zPriority = g_unkGraphicsStruct.g_zEntityCenter - 0x14;
-        entity->drawMode = g_eBreakableDrawModes[breakableType];
-        entity->hitboxHeight = g_eBreakableHitboxes[breakableType];
-        entity->animSet = g_eBreakableanimSets[breakableType];
-    }
-}
