+++ src/st/mad/e_breakable.c
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "mad.h"
+#include "sfx.h"
+
+#include "../e_breakable.h"
