+++ src/st/np3/np3.h
@@ -111,7 +111,6 @@ extern u16 g_EInitMerman[];
 extern u16 g_EInitMermanFireball[];
 extern u16 D_80180AF0[];
 extern u16 g_EInitBat[];
-extern u16 g_EInitBat[];
 extern u16 g_EInitZombie[];
 extern u16 D_80180B20[];
 extern u16 g_EInitSlograSpearNP3[];
@@ -220,8 +219,6 @@ extern u8 D_80182414[];
 extern s32 D_80182440[];
 extern u8 D_80182454[];
 extern u16 D_80182458[];
-extern u8 D_80182554[];
-extern u8 D_80182570[];
 extern u8 D_8018258C[];
 extern u8 D_80182594[];
 extern u8 D_801825A8[];
